requirements.txt
anthropic>=0.95.0
python-dotenv>=1.2.0
pydantic>=2.0.0

